requirements.txt
langchain==0.1.16
langchain-community==0.0.32

faiss-cpu==1.13.2
transformers>=4.36
huggingface-hub>=0.34

pypdf==4.0.1
streamlit==1.31.1
ollama==0.1.7
langchain-groq
python-dotenv

